+++ stingraykit/collection/MapDictionary.h
@@ -84,7 +84,7 @@ namespace stingray
 		MapDictionary& operator =(const MapDictionary& other)
 		{ _map.reset(new MapType(*other._map)); return *this; }
 
-		virtual int GetCount() const { return _map->size(); }
+		virtual size_t GetCount() const { return _map->size(); }
 		virtual bool IsEmpty() const { return _map->empty(); }
 
 		virtual bool TryGet(const KeyType& key, ValueType& outValue) const
